Move check inherit testing code to separate file
